+++ Starfire/utils/spritesheet.py
