+++ Starfire/view.py
@@ -1,17 +1,19 @@
 import pygame, sys
 from pygame.locals import *
-import model
-import background
-import spritesheet
+import Starfire.model as model
+import Starfire.utils.background as background
+
 
 import random
+
+BG_FILE='images/starfield800.bmp'
 
 class View:
 
    def __init__(self,screen_width,screen_height):
       
       self.screen = pygame.display.set_mode((screen_width, screen_height))
-      self.background = background.DownScrollingBackground(background.BG_FILE)
+      self.background = background.DownScrollingBackground(BG_FILE)
    
      
    def updateDisplay(self):
